backend/requirements.txt
# 核心框架
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.0
pydantic-settings

# Agno AI框架
agno==2.0.0

# 数据库 - 使用原生SQL驱动
aiosqlite==0.20.0  # SQLite异步驱动

# MCP相关
fastmcp==0.1.0

# Redis（用于任务存储）
redis==5.0.4

# 工具类
python-dotenv==1.0.1
loguru==0.7.2